Crate ink_e2e

Module for the logic behind ink!’s End-to-End testing framework.

Re-exports

pub use subxt;

pub use subxt_signer;

pub use tokio;

pub use tracing;

pub use tracing_subscriber;

Modules

events

preset

Exposes preset sandbox configurations to be used in tests.

sr25519

An sr25519 keypair implementation.

Structs

CallBuilder

Allows to build an end-to-end call using a builder pattern.

CallDryRunResult

Result of the dry run of a contract call.

CallResult

Result of a contract call.

CallTrace

A smart contract execution call trace.

Client

The Client takes care of communicating with the node.

ContractsRegistry

Registry of contracts that can be loaded.

DefaultSandbox

InstantiateBuilder

Allows to build an end-to-end instantiation call using a builder pattern.

InstantiateDryRunResult

Result of the dry run of a contract call.

InstantiationResult

Result of a contract instantiation.

Keypair

An sr25519 keypair implementation. While the API is slightly different, the logic for this has been taken from sp_core::sr25519 and we test against this to ensure conformity.

RpcClient

A concrete wrapper around an [RpcClientT] which provides some higher level helper methods and is cheaply cloneable.

SandboxClient

TestNodeProcess

Spawn a local substrate node for testing.

TestNodeProcessBuilder

Construct a test node process.

UploadResult

Result of a contract upload.

Weight

Enums

PolkadotConfig

Default set of commonly used types by Polkadot nodes.

Sr25519Keyring

Set of test accounts.

Constants

LOG_PREFIX

This prefix will be used for log output. It is set by each #[ink_e2e::test] with the function name as String. This way it is possible to distinguish the lines in stdout and stderr, to still know which line belongs to which test.

Statics

INIT

We use this to only initialize env_logger once.

Traits

BuilderClient

ChainBackend

General chain operations useful in contract testing.

ContractsBackend

Contract-specific operations.

E2EBackend

Full E2E testing backend: combines general chain API and contract-specific operations.

Functions

account_id

Get an ink! [ink_primitives::AccountId] for a given keyring account.

address

Returns the [ink::Address] for a given keyring account.

address_from_account_id

Returns the [ink::Address] for a given account id.

address_from_keypair

Returns the [ink::Address] for a given Keypair.

alice

Equivalent to {DEV_PHRASE}//Alice.

bob

Equivalent to {DEV_PHRASE}//Bob.

build_root_and_contract_dependencies

Builds the “root” contract (the contract in which the E2E tests are defined) together with any contracts which are a dependency of the root contract.

charlie

Equivalent to {DEV_PHRASE}//Charlie.

create_call_builder

Creates a call builder for Contract, based on an account id.

create_call_builder_abi

Creates a call builder for Contract for the specified ABI, based on an account id.

dave

Equivalent to {DEV_PHRASE}//Dave.

eve

Equivalent to {DEV_PHRASE}//Eve.

ferdie

Equivalent to {DEV_PHRASE}//Ferdie.

keypair_to_account

Transforms a Keypair into an account id.

log_error

Writes msg to stderr.

log_info

Writes msg to stdout.

log_prefix

Returns the name of the test which is currently executed.

one

Equivalent to {DEV_PHRASE}//One.

two

Equivalent to {DEV_PHRASE}//Two.

Type Aliases

CallBuilderFinal

Represents an initialized contract message builder.

Error

Attribute Macros

test

Defines an End-to-End test.