Function ink_e2e::build_message

pub fn build_message<Ref>(
    account_id: <DefaultEnvironment as Environment>::AccountId
) -> MessageBuilder<DefaultEnvironment, Ref>where
    Ref: TraitCallBuilder + FromAccountId<DefaultEnvironment>,

Convenience method for building messages for the default environment.

Note

This is hardcoded to ink_env::DefaultEnvironment so the user does not have to specify this generic parameter, which currently is hardcoded in the E2E testing suite.