Trait ink_e2e::ChainBackend

pub trait ChainBackend {
    type Actor: Send;
    type ActorId;
    type Balance: Send;
    type Error;
    type EventLog;

    // Required methods
    fn create_and_fund_account<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        origin: &'life1 Self::Actor,
        amount: Self::Balance
    ) -> Pin<Box<dyn Future<Output = Self::Actor> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn balance<'life0, 'async_trait>(
        &'life0 self,
        actor: Self::ActorId
    ) -> Pin<Box<dyn Future<Output = Result<Self::Balance, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn runtime_call<'a, 'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        actor: &'life1 Self::Actor,
        pallet_name: &'a str,
        call_name: &'a str,
        call_data: Vec<Value>
    ) -> Pin<Box<dyn Future<Output = Result<Self::EventLog, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

General chain operations useful in contract testing.

Required Associated Types

type Actor: Send

Abstract type representing the entity that interacts with the chain.

type ActorId

Identifier type for an actor.

type Balance: Send

Balance type.

type Error

Error type.

type EventLog

Event log type.

Required Methods

fn create_and_fund_account<'life0, 'life1, 'async_trait>( &'life0 mut self, origin: &'life1 Self::Actor, amount: Self::Balance ) -> Pin<Box<dyn Future<Output = Self::Actor> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Generate a new actor’s credentials and fund it with the given amount from the sender actor.

fn balance<'life0, 'async_trait>( &'life0 self, actor: Self::ActorId ) -> Pin<Box<dyn Future<Output = Result<Self::Balance, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the balance of actor.

fn runtime_call<'a, 'life0, 'life1, 'async_trait>( &'life0 mut self, actor: &'life1 Self::Actor, pallet_name: &'a str, call_name: &'a str, call_data: Vec<Value> ) -> Pin<Box<dyn Future<Output = Result<Self::EventLog, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Executes a runtime call call_name for the pallet_name. The call_data is a Vec<Value>.

Note:

pallet_name must be in camel case, for example Balances.

call_name must be snake case, for example force_transfer.

call_data is a Vec<subxt::dynamic::Value> that holds a representation of some value.

Returns when the transaction is included in a block. The return value contains all events that are associated with this transaction.