Trait ink_e2e::ContractsBackend

pub trait ContractsBackend<E: Environment> {
    type Actor;
    type Error;
    type EventLog;

    // Required methods
    fn instantiate<'life0, 'life1, 'life2, 'async_trait, Contract, Args, R>(
        &'life0 mut self,
        contract_name: &'life1 str,
        caller: &'life2 Self::Actor,
        constructor: CreateBuilder<E, Contract, Unset<<E as Environment>::Hash>, Unset<u64>, Unset<<E as Environment>::Balance>, Set<ExecutionInput<Args>>, Unset<Salt>, Set<ReturnType<R>>>,
        value: E::Balance,
        storage_deposit_limit: Option<E::Balance>
    ) -> Pin<Box<dyn Future<Output = Result<InstantiationResult<E, Self::EventLog>, Self::Error>> + Send + 'async_trait>>
       where Contract: 'async_trait,
             Args: 'async_trait + Send + Encode,
             R: 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn instantiate_dry_run<'life0, 'life1, 'life2, 'async_trait, Contract, Args, R>(
        &'life0 mut self,
        contract_name: &'life1 str,
        caller: &'life2 Self::Actor,
        constructor: CreateBuilder<E, Contract, Unset<<E as Environment>::Hash>, Unset<u64>, Unset<<E as Environment>::Balance>, Set<ExecutionInput<Args>>, Unset<Salt>, Set<ReturnType<R>>>,
        value: E::Balance,
        storage_deposit_limit: Option<E::Balance>
    ) -> Pin<Box<dyn Future<Output = ContractInstantiateResult<E::AccountId, E::Balance, ()>> + Send + 'async_trait>>
       where Contract: 'async_trait,
             Args: 'async_trait + Send + Encode,
             R: 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn upload<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        contract_name: &'life1 str,
        caller: &'life2 Self::Actor,
        storage_deposit_limit: Option<E::Balance>
    ) -> Pin<Box<dyn Future<Output = Result<UploadResult<E, Self::EventLog>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn call<'life0, 'life1, 'life2, 'async_trait, Args, RetType>(
        &'life0 mut self,
        caller: &'life1 Self::Actor,
        message: &'life2 CallBuilderFinal<E, Args, RetType>,
        value: E::Balance,
        storage_deposit_limit: Option<E::Balance>
    ) -> Pin<Box<dyn Future<Output = Result<CallResult<E, RetType, Self::EventLog>, Self::Error>> + Send + 'async_trait>>
       where CallBuilderFinal<E, Args, RetType>: Clone,
             Args: 'async_trait + Sync + Encode,
             RetType: 'async_trait + Send + Decode,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn call_dry_run<'life0, 'life1, 'life2, 'async_trait, Args, RetType>(
        &'life0 mut self,
        caller: &'life1 Self::Actor,
        message: &'life2 CallBuilderFinal<E, Args, RetType>,
        value: E::Balance,
        storage_deposit_limit: Option<E::Balance>
    ) -> Pin<Box<dyn Future<Output = CallDryRunResult<E, RetType>> + Send + 'async_trait>>
       where CallBuilderFinal<E, Args, RetType>: Clone,
             Args: 'async_trait + Sync + Encode,
             RetType: 'async_trait + Send + Decode,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Contract-specific operations.

Required Associated Types

type Actor

Abstract type representing the entity that interacts with the chain.

type Error

Error type.

type EventLog

Event log type.

Required Methods

fn instantiate<'life0, 'life1, 'life2, 'async_trait, Contract, Args, R>( &'life0 mut self, contract_name: &'life1 str, caller: &'life2 Self::Actor, constructor: CreateBuilder<E, Contract, Unset<<E as Environment>::Hash>, Unset<u64>, Unset<<E as Environment>::Balance>, Set<ExecutionInput<Args>>, Unset<Salt>, Set<ReturnType<R>>>, value: E::Balance, storage_deposit_limit: Option<E::Balance> ) -> Pin<Box<dyn Future<Output = Result<InstantiationResult<E, Self::EventLog>, Self::Error>> + Send + 'async_trait>>where Contract: 'async_trait, Args: 'async_trait + Send + Encode, R: 'async_trait, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

The function subsequently uploads and instantiates an instance of the contract.

This function extracts the metadata of the contract at the file path target/ink/$contract_name.contract.

Calling this function multiple times should be idempotent, the contract is newly instantiated each time using a unique salt. No existing contract instance is reused!

fn instantiate_dry_run<'life0, 'life1, 'life2, 'async_trait, Contract, Args, R>( &'life0 mut self, contract_name: &'life1 str, caller: &'life2 Self::Actor, constructor: CreateBuilder<E, Contract, Unset<<E as Environment>::Hash>, Unset<u64>, Unset<<E as Environment>::Balance>, Set<ExecutionInput<Args>>, Unset<Salt>, Set<ReturnType<R>>>, value: E::Balance, storage_deposit_limit: Option<E::Balance> ) -> Pin<Box<dyn Future<Output = ContractInstantiateResult<E::AccountId, E::Balance, ()>> + Send + 'async_trait>>where Contract: 'async_trait, Args: 'async_trait + Send + Encode, R: 'async_trait, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Dry run contract instantiation.

fn upload<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, contract_name: &'life1 str, caller: &'life2 Self::Actor, storage_deposit_limit: Option<E::Balance> ) -> Pin<Box<dyn Future<Output = Result<UploadResult<E, Self::EventLog>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

The function subsequently uploads and instantiates an instance of the contract.

This function extracts the Wasm of the contract for the specified contract.

Calling this function multiple times should be idempotent, the contract is newly instantiated each time using a unique salt. No existing contract instance is reused!

fn call<'life0, 'life1, 'life2, 'async_trait, Args, RetType>( &'life0 mut self, caller: &'life1 Self::Actor, message: &'life2 CallBuilderFinal<E, Args, RetType>, value: E::Balance, storage_deposit_limit: Option<E::Balance> ) -> Pin<Box<dyn Future<Output = Result<CallResult<E, RetType, Self::EventLog>, Self::Error>> + Send + 'async_trait>>where CallBuilderFinal<E, Args, RetType>: Clone, Args: 'async_trait + Sync + Encode, RetType: 'async_trait + Send + Decode, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Executes a call for the contract at account_id.

Returns when the transaction is included in a block. The return value contains all events that are associated with this transaction.

fn call_dry_run<'life0, 'life1, 'life2, 'async_trait, Args, RetType>( &'life0 mut self, caller: &'life1 Self::Actor, message: &'life2 CallBuilderFinal<E, Args, RetType>, value: E::Balance, storage_deposit_limit: Option<E::Balance> ) -> Pin<Box<dyn Future<Output = CallDryRunResult<E, RetType>> + Send + 'async_trait>>where CallBuilderFinal<E, Args, RetType>: Clone, Args: 'async_trait + Sync + Encode, RetType: 'async_trait + Send + Decode, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Executes a dry-run call.

Returns the result of the dry run, together with the decoded return value of the invoked message.