Crate ink_e2e

Module for the logic behind ink!’s End-to-End testing framework.

Re-exports

pub use subxt;

pub use tokio;

pub use tracing_subscriber;

Modules

events

sr25519

An sr25519 keypair implementation.

Macros

build

Builds a contract and imports its scaffolded structure as a module.

Structs

CallBuilder

Allows to build an end-to-end call using a builder pattern.

CallDryRunResult

Result of the dry run of a contract call.

CallResult

Result of a contract call.

Client

The Client takes care of communicating with the node.

DrinkClient

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

InstantiateBuilder

Allows to build an end-to-end instantiation call using a builder pattern.

InstantiateDryRunResult

Result of the dry run of a contract call.

InstantiationResult

Result of a contract instantiation.

Keypair

An sr25519 keypair implementation. While the API is slightly different, the logic for this has been taken from sp_core::sr25519 and we test against this to ensure conformity.

MinimalRuntime

RpcClient

A concrete wrapper around an [RpcClientT] which provides some higher level helper methods, is cheaply cloneable, and can be handed to things like [crate::client::OnlineClient] to instantiate it.

TestNodeProcess

Spawn a local substrate node for testing.

TestNodeProcessBuilder

Construct a test node process.

UploadResult

Result of a contract upload.

Enums

AccountKeyring

Set of test accounts.

PolkadotConfig

Default set of commonly used types by Polkadot nodes.

Constants

LOG_PREFIX

This prefix will be used for log output. It is set by each #[ink_e2e::test] with the function name as String. This way it is possible to distinguish the lines in stdout and stderr, to still know which line belongs to which test.

Statics

INIT

We use this to only initialize env_logger once.

Traits

ChainBackend

General chain operations useful in contract testing.

ContractsBackend

Contract-specific operations.

E2EBackend

Full E2E testing backend: combines general chain API and contract-specific operations.

Functions

account_id

Get an ink! ink_primitives::AccountId for a given keyring account.

alice

Equivalent to {DEV_PHRASE}//Alice.

bob

Equivalent to {DEV_PHRASE}//Bob.

build_root_and_additional_contracts

Builds the “root” contract (the contract in which the E2E tests are defined) together with the additional contracts specified in the additional_contracts argument.

build_root_and_contract_dependencies

Builds the “root” contract (the contract in which the E2E tests are defined) together with any contracts which are a dependency of the root contract.

charlie

Equivalent to {DEV_PHRASE}//Charlie.

dave

Equivalent to {DEV_PHRASE}//Dave.

eve

Equivalent to {DEV_PHRASE}//Eve.

ferdie

Equivalent to {DEV_PHRASE}//Ferdie.

log_error

Writes msg to stderr.

log_info

Writes msg to stdout.

log_prefix

Returns the name of the test which is currently executed.

one

Equivalent to {DEV_PHRASE}//One.

two

Equivalent to {DEV_PHRASE}//Two.

Type Definitions

CallBuilderFinal

Represents an initialized contract message builder.

Error

Attribute Macros

test

Defines an End-to-End test.