Module ink_e2e::preset::mock_network::mocks::msg_queue::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

[Pallet], which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the [dispatchables] module or in the [Call] enum.

[storage_types], which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

[Config], which contains the configuration trait of this pallet.

[Event] and [Error], which are listed among the Enums.

Structs

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Type Aliases

MessageId

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.