ink_e2e::preset::mock_network::mocks::msg_queue::pallet
Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type XcmExecutor: ExecuteXcm<Self::RuntimeCall>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

type XcmExecutor: ExecuteXcm<Self::RuntimeCall>

Object Safety

This trait is not object safe.