Module ink_e2e::sr25519

An sr25519 keypair implementation.

Modules

dev

Dev accounts, helpful for testing but not to be used in production, since the secret keys are known.

Structs

Keypair

An sr25519 keypair implementation. While the API is slightly different, the logic for this has been taken from sp_core::sr25519 and we test against this to ensure conformity.

PublicKey

The public key for an Keypair key pair. This is equivalent to a Substrate AccountId32.

Signature

A signature generated by Keypair::sign(). These bytes are equivalent to a Substrate MultiSignature::sr25519(bytes).

Enums

Error

An error handed back if creating a keypair fails.

Functions

verify

Verify that some signature for a message was created by the owner of the PublicKey.

Type Aliases

Seed

Seed bytes used to generate a key pair.