Module ink_e2e::sr25519::dev

Dev accounts, helpful for testing but not to be used in production, since the secret keys are known.

Functions

alice

Equivalent to {DEV_PHRASE}//Alice.

bob

Equivalent to {DEV_PHRASE}//Bob.

charlie

Equivalent to {DEV_PHRASE}//Charlie.

dave

Equivalent to {DEV_PHRASE}//Dave.

eve

Equivalent to {DEV_PHRASE}//Eve.

ferdie

Equivalent to {DEV_PHRASE}//Ferdie.

one

Equivalent to {DEV_PHRASE}//One.

two

Equivalent to {DEV_PHRASE}//Two.